requirements.txt
audiofile==1.3.2
audioread==3.0.1
huggingface-hub==0.19.4
librosa==0.10.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
numpy==1.26.2
onnx==1.15.0
onnx-pytorch==0.1.5
onnxruntime==1.16.3
optimum==1.17.1
pandas==2.1.3
PyAudio==0.2.11
pytorch-transformers==1.2.0
soundfile==0.12.1
torch==2.1.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.1.1